propagate the warning to other __WARN__ handlers
For __DIE__ handlers, simply calling die($msg) is sufficient for other __DIE__
handlers to see the message, but __WARN__ handlers are not called when calling
warn() inside one, so we need to do this explicitly.  One observable side
effect of this change is that Devel::Confess will be able to provide stack
trades for warnings when Test::Warnings is also loaded.