fix warnings->import being called after importing "warnings" sub
If the user imports the "warnings" sub, it will take priority over the
bareword 'warnings'.  This means that a manually written
"warnings->import" call will try to call the warnings sub from this
module.  One example of this is older versions of File::pushd, which
conditionally called warnings->import before setting their package, so
it will instead happen in whatever package the module was loaded from.

We can mark the sub argument as optional, which shouldn't ever normally
happen except for when trying to call a method on warnings.  Then we can
check the number of arguments to detect this, and return the string
'warnings', allowing method lookup to proceed like normal.