Add :report_warnings functionality
If a test uses Test::Output::stderr_like for example, which captures
STDERR, any additional warning in that block is not seen in the output,
but is reported correctly by Test::Warnings.
So it requires manual changes to find out what the warning was.

Wrapping an explicit `warnings {}` around those calls and printing
the warnings helps, but is not really usable when there are many
tests like this.

This change reports the actual contents of those warnings at the end (if
requested).

I still need to find out how to make the new test not fail, and
also replace Test::Output with a simple STDERR redirection.